+++ lib/domain_name_format_validator.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "domain_name_format_validator/settings"
 require "domain_name_format_validator/validator"
 require "domain_name_format_validator/version"